requirements_fixed.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-pinecone>=0.1.0
huggingface-hub>=0.30.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
transformers>=4.30.0
torch>=2.0.0 